+++ packages/core/src/components/Logo/Logo.styled.tsx
@@ -1,0 +1,8 @@
+import { styled } from '@medly-components/utils';
+
+export const Container = styled.div`
+    display: inline-grid;
+    grid-auto-flow: column;
+    grid-column-gap: 0.8rem;
+    align-items: center;
+`